refactor(api/user): change User entity namespace
