+++ modules/back-end/src/Infrastructure/MongoDb/MongoDbClient.cs
@@ -1,4 +1,4 @@
-using Domain.Identity;
+using Domain.Users;
 using Microsoft.Extensions.Options;
 using MongoDB.Driver;
 using MongoDB.Driver.Linq;
