+++ modules/back-end/tests/Application.IntegrationTests/Stubs/TestPasswordHasher.cs
@@ -1,4 +1,4 @@
-using Domain.Identity;
+using Domain.Users;
 using Microsoft.AspNetCore.Identity;
 
 namespace Application.IntegrationTests.Stubs;
