+++ modules/back-end/tests/Application.UnitTests/Identity/IdentityServiceTests.cs
@@ -1,4 +1,4 @@
-using Domain.Identity;
+using Domain.Users;
 using Infrastructure.Identity;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Options;
